requirements.txt
discord
yfinance
requests
stockstats
python-dotenv
schedule